requirements.txt
aniso8601==9.0.1
attrs==21.2.0
certifi==2020.12.5
chardet==4.0.0
click==8.0.0
Flask==2.0.0
Flask-HTTPAuth==4.4.0
greenlet==1.1.0
idna==2.10
iniconfig==1.1.1
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
packaging==20.9
pluggy==0.13.1
py==1.10.0
pyparsing==2.4.7
pytest==6.2.4
python-decouple==3.4
pytz==2021.1
requests==2.25.1
six==1.16.0
toml==0.10.2
urllib3==1.26.4
Werkzeug==2.0.1
